Update Two_sum.py
This is a code where from a given list of numbers we want to select the indices of two numbers whose sum is a given target number. Now there are several ways to do it, here we present one simple method which uses two important concepts dictionary and list. We will see the syntax to use these .